👌 extract command.js
