+++ scripts/deploy/deploy.js
@@ -1,6 +1,7 @@
 'use strict'
 
-const { printLog, command, runMain } = require('../lib/execution-utils')
+const { printLog, runMain } = require('../lib/execution-utils')
+const { command } = require('../lib/command')
 const {
   buildRootUploadPath,
   buildDatacenterUploadPath,
