+++ scripts/test/bump-chrome-driver-version.js
@@ -1,7 +1,8 @@
 'use strict'
 
 const fs = require('fs')
-const { printLog, printError, runMain, command, fetch } = require('../lib/execution-utils')
+const { printLog, printError, runMain, fetch } = require('../lib/execution-utils')
+const { command } = require('../lib/command')
 const { CI_FILE, replaceCiFileVariable } = require('../lib/files-utils')
 const { initGitConfig } = require('../lib/git-utils')
 const { getGithubAccessToken } = require('../lib/secrets')
