+++ scripts/test/export-test-result.js
@@ -1,6 +1,7 @@
 'use strict'
 
-const { command, printLog, runMain } = require('../lib/execution-utils')
+const { printLog, runMain } = require('../lib/execution-utils')
+const { command } = require('../lib/command')
 const { getOrg2ApiKey } = require('../lib/secrets')
 
 /**
